feat: abstract search components for reusabilityg
